feat: expose configs always
