+++ core/src/services/atomicserver/mod.rs
@@ -18,4 +18,3 @@
 mod backend;
 
 pub use backend::AtomicserverBuilder as Atomicserver;
-pub use backend::AtomicserverConfig;
